// meteor/client/lib/forms/schema-types.ts
export const $schema = 'https://json-schema.org/draft/2020-12/schema' as const

type MaybeReadonlyArray<T> = Array<T> | ReadonlyArray<T>
type ValueOf<T> = T[keyof T]
// ...
export namespace JSONSchema {
	export type TypeValue =
		| ValueOf<TypeName>
		| TypeName
		| Array<ValueOf<TypeName> | TypeName>
		| ReadonlyArray<ValueOf<TypeName> | TypeName>

	/**
	 * JSON Schema interface
	 */
	export type Interface<Value = any, SchemaType extends TypeValue = TypeValue> = Exclude<
		JSONSchema<Value, SchemaType>,
		boolean
	>

	export type Array<T = any> = Pick<Interface<T, 'array'>, KeywordByType.Any | KeywordByType.Array>

	export type Boolean = Pick<Interface<boolean, 'boolean'>, KeywordByType.Any>

	export type Integer = Pick<Interface<number, 'integer'>, KeywordByType.Any | KeywordByType.Number>

	export type Number = Pick<Interface<number, 'number'>, KeywordByType.Any | KeywordByType.Number>

	export type Null = Pick<Interface<null, 'null'>, KeywordByType.Any>

	export type Object<T = any> = Pick<Interface<T, 'object'>, KeywordByType.Any | KeywordByType.Object>

	export type String = Pick<Interface<string, 'string'>, KeywordByType.Any | KeywordByType.String>
}

namespace KeywordByType {
	export type Any =
		| '$anchor'
		| '$comment'
		| '$defs'
		| '$dynamicAnchor'
		| '$dynamicRef'
		| '$id'
		| '$ref'
		| '$schema'
		| '$vocabulary'
		| 'allOf'
		| 'anyOf'
		| 'const'
		| 'default'
		| 'definitions'
		| 'deprecated'
		| 'description'
		| 'else'
		| 'enum'
		| 'examples'
		| 'format'
		| 'if'
		| 'not'
		| 'oneOf'
		| 'readOnly'
		| 'then'
		| 'title'
		| 'type'
		| 'writeOnly'

	export type Array =
		| 'additionalItems'
		| 'contains'
		| 'items'
		| 'maxContains'
		| 'maxItems'
		| 'minContains'
		| 'minItems'
		| 'prefixItems'
		| 'unevaluatedItems'
		| 'uniqueItems'

	export type Number = 'exclusiveMaximum' | 'exclusiveMinimum' | 'maximum' | 'minimum' | 'multipleOf'

	export type Object =
		| 'additionalProperties'
		| 'dependencies'
		| 'dependentRequired'
		| 'dependentSchemas'
		| 'maxProperties'
		| 'minProperties'
		| 'patternProperties'
		| 'properties'
		| 'propertyNames'
		| 'required'
		| 'unevaluatedProperties'

	export type String =
		| 'contentEncoding'
		| 'contentMediaType'
		| 'contentSchema'
		| 'maxLength'
		| 'minLength'
		| 'pattern'
}
// ...
export enum TypeName {
	/**
	 * Value MUST be an array.
	 */
	Array = 'array',

	/**
	 * Value MUST be a boolean.
	 */
	Boolean = 'boolean',

	/**
	 * Value MUST be an integer, no floating point numbers are allowed. This is a
	 * subset of the number type.
	 */
	Integer = 'integer',

	/**
	 * Value MUST be null. Note this is mainly for purpose of being able use union
	 * types to define nullability. If this type is not included in a union, null
	 * values are not allowed (the primitives listed above do not allow nulls on
	 * their own).
	 */
	Null = 'null',

	/**
	 * Value MUST be a number, floating point numbers are allowed.
	 */
	Number = 'number',

	/**
	 * Value MUST be an object.
	 */
	Object = 'object',

	/**
	 * Value MUST be a string.
	 */
	String = 'string',
}